requirements.txt
amqp==5.1.1
asgiref==3.3.4
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==5.5
Django==3.2
django-cors-headers==3.7.0
django-environ==0.8.1
djangorestframework==3.12.4
gunicorn==20.1.0
kombu==5.2.4
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pytz==2022.1
six==1.16.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
